pyproject.toml
[tool.poetry]
name = "coinbase-dashboard-streamlit"
version = "0.1.0"
description = ""
authors = ["Eliminate-Entrant <mshaheerch@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.34.0"
black = "^24.4.2"
pytest = "^8.2.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
plotly = "^5.22.0"
pandas = "^2.2.2"
websockets = "^12.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
